# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ExceptionCreate": {
        "additionalProperties": false,
        "properties": {
          "application": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Application",
            "type": "string"
          },
          "business_justification": {
            "maxLength": 4000,
            "minLength": 20,
            "title": "Business Justification",
            "type": "string"
          },
          "compensating_controls": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Compensating Controls",
            "type": "array"
          },
          "control_id": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Control Id",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "granularity": {
            "enum": [
              "RESOURCE",
              "SCOPE",
              "PRINCIPAL"
            ],
            "title": "Granularity",
            "type": "string"
          },
          "implementation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Implementation Id"
          },
          "principal_patterns": {
            "items": {
              "type": "string"
            },
            "maxItems": 10,
            "title": "Principal Patterns",
            "type": "array"
          },
          "resource_ids": {
            "items": {
              "type": "string"
            },
            "maxItems": 50,
            "title": "Resource Ids",
            "type": "array"
          },
          "risk_owner": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Risk Owner",
            "type": "string"
          },
          "scope_id": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Scope Id",
            "type": "string"
          },
          "team": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Team",
            "type": "string"
          },
          "technical_justification": {
            "maxLength": 4000,
            "minLength": 20,
            "title": "Technical Justification",
            "type": "string"
          },
          "valid_from": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Valid From"
          },
          "work_ref": {
            "anyOf": [
              {
                "maxLength": 256,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Work Ref"
          }
        },
        "required": [
          "control_id",
          "application",
          "team",
          "business_justification",
          "technical_justification",
          "scope_id",
          "granularity",
          "risk_owner",
          "compensating_controls",
          "expires_at"
        ],
        "title": "ExceptionCreate",
        "type": "object"
      },
      "ExceptionDecision": {
        "additionalProperties": false,
        "properties": {
          "decision": {
            "enum": [
              "APPROVE",
              "REJECT"
            ],
            "title": "Decision",
            "type": "string"
          },
          "expected_lock_version": {
            "title": "Expected Lock Version",
            "type": "integer"
          },
          "rationale": {
            "maxLength": 4000,
            "minLength": 10,
            "title": "Rationale",
            "type": "string"
          }
        },
        "required": [
          "decision",
          "rationale",
          "expected_lock_version"
        ],
        "title": "ExceptionDecision",
        "type": "object"
      },
      "ExceptionRenewal": {
        "additionalProperties": false,
        "properties": {
          "business_justification": {
            "maxLength": 4000,
            "minLength": 20,
            "title": "Business Justification",
            "type": "string"
          },
          "compensating_controls": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "minItems": 1,
            "title": "Compensating Controls",
            "type": "array"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "technical_justification": {
            "maxLength": 4000,
            "minLength": 20,
            "title": "Technical Justification",
            "type": "string"
          }
        },
        "required": [
          "expires_at",
          "business_justification",
          "technical_justification",
          "compensating_controls"
        ],
        "title": "ExceptionRenewal",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LockedCommand": {
        "additionalProperties": false,
        "properties": {
          "expected_lock_version": {
            "title": "Expected Lock Version",
            "type": "integer"
          }
        },
        "required": [
          "expected_lock_version"
        ],
        "title": "LockedCommand",
        "type": "object"
      },
...
      "PlanStateCommand": {
        "additionalProperties": false,
        "properties": {
          "expected_lock_version": {
            "title": "Expected Lock Version",
            "type": "integer"
          },
          "reason": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason",
          "expected_lock_version"
        ],
        "title": "PlanStateCommand",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Security intent -> governed, verifiable preventive controls. Fixture data only; no cloud mutations. Delivery is a local GitOps handoff bundle and a clearly labelled mock pipeline.",
    "title": "Cloud Security Control Engineering Platform (local MVP)",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/exceptions": {
      "get": {
        "operationId": "list_exceptions_api_v1_exceptions_get",
        "parameters": [
          {
            "in": "query",
            "name": "control_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Control Id"
            }
          },
          {
            "in": "query",
            "name": "effective_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Effective Status"
            }
          },
          {
            "in": "query",
            "name": "disposition",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Disposition"
            }
          },
          {
            "in": "query",
            "name": "expiring",
            "required": false,
            "schema": {
              "default": false,
              "title": "Expiring",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "native_cleanup_pending",
            "required": false,
            "schema": {
              "default": false,
              "title": "Native Cleanup Pending",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "maximum": 100000,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Exceptions",
        "tags": [
          "exceptions"
        ]
      },
      "post": {
        "operationId": "request_exception_api_v1_exceptions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExceptionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Exception",
        "tags": [
          "exceptions"
        ]
      }
    },
    "/api/v1/exceptions/{exc_id}": {
      "get": {
        "operationId": "get_exception_api_v1_exceptions__exc_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "exc_id",
            "required": true,
            "schema": {
              "title": "Exc Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Exception",
        "tags": [
          "exceptions"
        ]
      }
    },
    "/api/v1/exceptions/{exc_id}/decision": {
      "post": {
        "operationId": "decide_api_v1_exceptions__exc_id__decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "exc_id",
            "required": true,
            "schema": {
              "title": "Exc Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExceptionDecision"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide",
        "tags": [
          "exceptions"
        ]
      }
    },
    "/api/v1/exceptions/{exc_id}/renew": {
      "post": {
        "operationId": "renew_api_v1_exceptions__exc_id__renew_post",
        "parameters": [
          {
            "in": "path",
            "name": "exc_id",
            "required": true,
            "schema": {
              "title": "Exc Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExceptionRenewal"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Renew",
        "tags": [
          "exceptions"
        ]
      }
    },
    "/api/v1/exceptions/{exc_id}/revoke": {
      "post": {
        "operationId": "revoke_api_v1_exceptions__exc_id__revoke_post",
        "parameters": [
          {
            "in": "path",
            "name": "exc_id",
            "required": true,
            "schema": {
              "title": "Exc Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanStateCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke",
        "tags": [
          "exceptions"
        ]
      }
    },
    "/api/v1/exceptions/{exc_id}/start-review": {
      "post": {
        "operationId": "start_review_api_v1_exceptions__exc_id__start_review_post",
        "parameters": [
          {
            "in": "path",
            "name": "exc_id",
            "required": true,
            "schema": {
              "title": "Exc Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LockedCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Review",
        "tags": [
          "exceptions"
        ]
      }
    },
...
  }
}